first pass at moving files methods into FilesProvider
