+++ dbt_platform_helper/utils/files.py
@@ -1,32 +1,13 @@
-from os import makedirs
-from pathlib import Path
-
 import click
 import yaml
 from jinja2 import Environment
 from jinja2 import FileSystemLoader
 
+from dbt_platform_helper.providers.files import FileProvider
+
 
 def to_yaml(value):
     return yaml.dump(value, sort_keys=False)
-
-
-def mkfile(base_path, file_path, contents, overwrite=False):
-    file_path = Path(file_path)
-    file = Path(base_path).joinpath(file_path)
-    file_exists = file.exists()
-
-    if not file_path.parent.exists():
-        makedirs(file_path.parent)
-
-    if file_exists and not overwrite:
-        return f"File {file_path} exists; doing nothing"
-
-    action = "overwritten" if file_exists and overwrite else "created"
-
-    file.write_text(contents)
-
-    return f"File {file_path} {action}"
 
 
 def generate_override_files(base_path, file_path, output_dir):
@@ -36,7 +17,7 @@ def generate_override_files(base_path, file_path, output_dir):
                 contents = file.read_text()
                 file_name = str(file).removeprefix(f"{file_path}/")
                 click.echo(
-                    mkfile(
+                    FileProvider.mkfile(
                         base_path,
                         output_dir / file_name,
                         contents,
@@ -61,7 +42,9 @@ def generate_override_files_from_template(base_path, overrides_path, output_dir,
             if file.is_file():
                 file_name = str(file).removeprefix(f"{overrides_path}/")
                 contents = templates.get_template(str(file_name)).render(data)
-                message = mkfile(base_path, output_dir / file_name, contents, overwrite=True)
+                message = FileProvider.mkfile(
+                    base_path, output_dir / file_name, contents, overwrite=True
+                )
                 click.echo(message)
 
     generate_files_for_dir("*")
